requirements.txt
asgiref==3.8.1
Django==4.2.12
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
setuptools==70.1.0
sqlparse==0.5.0
uritemplate==4.1.1
